Data Science Example
This is a study of vacancy rate's effect on homeless populations in the US. A self created panel data set was used in a multi-variate quadratic fixed effects regression, by state and year. A summary of the results can be read at the end of the analysis_output.do file. The full report with visualizations can be found in my repository as, "Data Science Project: Full Report".